requirements.txt
absl-py==0.13.0
-e git+git@github.com:nickjalbert/agentos.git@a9dcdde74ddcde7ed344300fc165c2720de024ad#egg=agentos
alembic==1.6.5
appdirs==1.4.3
astunparse==1.6.3
azure-core==1.16.0
azure-storage-blob==12.8.1
black==21.6b0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
click==8.0.1
cloudpickle==1.3.0
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.7
cycler==0.10.0
databricks-cli==0.14.3
decorator==5.0.9
distlib==0.3.0
distro==1.4.0
dm-acme==0.2.1
dm-env==1.5
dm-reverb==0.3.1
dm-reverb-nightly==0.3.0.dev20210712
dm-sonnet==2.0.0
dm-tree==0.1.6
docker==5.0.0
EasyProcess==0.3
entrypoints==0.3
factories==1.2.0
flake8==3.9.2
Flask==2.0.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
gorilla==0.4.0
grpcio==1.34.1
gunicorn==20.1.0
gym==0.17.1
h5py==3.1.0
html5lib==1.0.1
idna==2.8
imageio==2.9.0
ipaddr==2.2.0
isodate==0.6.0
itsdangerous==2.0.1
jax==0.2.17
jaxlib==0.1.68
Jinja2==3.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
launchpad==1.1.0
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
mlflow==1.9.1
msgpack==0.6.2
msrest==0.6.21
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.3
pandas==1.3.0
pathspec==0.8.1
pep517==0.8.2
Pillow==8.3.1
portpicker==1.4.0
progress==1.5
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyglet==1.5.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-editor==1.0.4
pytoml==0.1.21
pytz==2021.1
PyVirtualDisplay==2.2
PyYAML==5.4.1
querystring-parser==1.2.4
qute==3.0.11
regex==2021.7.6
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scipy==1.7.0
scribble==1.0.0
six==1.15.0
smmap==4.0.0
SQLAlchemy==1.3.13
sqlparse==0.4.1
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-probability==0.13.0
termcolor==1.1.0
toml==0.10.2
trfl==1.1.0
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
wrapt==1.12.1
